refactor(frontend): view layer state
- move shared view layer state into `src/lib/`.
- document the data map functions that generate the deck.gl layers.
- fix a duplicate Recoil key for `viewLayerState`.
