+++ frontend/src/app/map/DataMap.tsx
@@ -1,9 +1,9 @@
-import { FC, useEffect } from 'react';
+import { FC } from 'react';
 import { useRecoilValue } from 'recoil';
 
 import { interactionGroupsState } from 'app/state/layers/interaction-groups';
 import { viewLayersFlatState } from 'app/state/layers/view-layers-flat';
-import { useSaveViewLayers, viewLayersParamsState } from 'app/state/layers/view-layers-params';
+import { viewLayersParamsState } from 'app/state/layers/view-layers-params';
 
 import { DataMap } from 'lib/data-map/DataMap';
 
@@ -14,15 +14,8 @@ export const DataMapContainer: FC = () => {
   const background = useRecoilValue(backgroundState);
   const showLabels = useRecoilValue(showLabelsState);
   const viewLayers = useRecoilValue(viewLayersFlatState);
-  const saveViewLayers = useSaveViewLayers();
   const { firstLabelId } = useBasemapStyle(background, showLabels);
-
-  useEffect(() => {
-    saveViewLayers(viewLayers);
-  }, [saveViewLayers, viewLayers]);
-
   const viewLayersParams = useRecoilValue(viewLayersParamsState);
-
   const interactionGroups = useRecoilValue(interactionGroupsState);
 
   return (
